+++ views/settings-form.php
@@ -7,7 +7,7 @@ if (!defined('ABSPATH')) {
 ?>
 
 <div class="wrap">
-	<h2><?php esc_html_e('Email Logs', 'log-emails'); ?></h2>
+	<h2><?php esc_html_e('Log Emails', 'log-emails'); ?></h2>
 
 	<form action="<?php echo admin_url('options.php'); ?>" method="POST">
 		<?php settings_fields('log-emails'); ?>